Move sanity test code into separate packages
